The original plan was to find all bind mounts, and bind mount the same way in relation to the layout that was being built with overlayfses as in a bind mounted /media/ -> /mnt  would be done with bindmounting ${TMP_DIR}/workdir/media -> ${TMP_DIR}/workdir/mnt to reduce the number of overlayfses created.
The way bind mounts are handled however, and actually stored in the kernel, if /dev/sda1 is root, and say /home is bind mounted to /users
/dev/sda1 is actually handled as being mounted again, only with the root being /home, so it is closer to saying such as mounting /dev/sda:home/ to /users

This also is how subvolumes of BTRFS volumes are mounted, and presented in /proc/self/mountinfo .

Its cleaner to just create a new overlayfs for each bind mount. However files can be bind mounted. overlayfs requires there to be directories, so if this is detected, fall back