Sync fixes.
Allow it to work on newer versions of Linux. The test for if the lowerdir was not working after changes

Speed up unneeded handling when OVERLAYFS translation is being used. breeze-icons containing thousands of items takes much shorter to complete

Empty directories in overlayfs translation can now be included in packages